- changes Vol data type to decimal since older msft data use decimal

+++ QuantTrading/QuantTrading.Domain/Models/MarketData.cs
@@ -7,4 +7,4 @@ public sealed record MarketData(
     decimal High,
     decimal Low,
     decimal Close,
-    long Volume);
+    decimal Volume);
